+++ main1139.cpp
@@ -28,7 +28,7 @@ int main() {
 	jyhnie7u9	=	(unsigned short)-334273957;
 	JD058e2	=	(unsigned short)-1146721985;
 	qOn3hb	=	(unsigned short)-572551953;
-	Tec84Bfcw	=	882317617;
+	Tec84Bfcw	=	(unsigned short)882317617;
 	qM4d	=	1630010055;
 	d9y7	=	875895319;
 	aVMZeK	=	484132526;
@@ -38,7 +38,7 @@ int main() {
 	V55OC_2Ebq	=	(unsigned short)-995803375;
 	v0jbiRiTi	=	(unsigned short)-9811487;
 	y6MNC5uY2	=	(unsigned short)-520829028;
-	e_DSS	=	687839690;
+	e_DSS	=	(unsigned short)687839690;
 	rb198R04	=	-1209714125;
 	ulWF72HS	=	98379649;
 	fa3f9	=	-1610517255;
@@ -48,7 +48,7 @@ int main() {
 	Hh1Fm4_	=	(unsigned short)2034122852;
 	RB_5	=	(unsigned short)-111721498;
 	a9A095fDiD	=	(unsigned short)49658445;
-	NkZw4mgp	=	1822202263;
+	NkZw4mgp	=	(unsigned short)1822202263;
 	V_0877	=	-379864093;
 	aPuGS_np2	=	-11475788;
 	PhY_	=	792060143;
@@ -58,7 +58,7 @@ int main() {
 	Gc_lFC69	=	(unsigned short)-1587994485;
 	UBQkK	=	(unsigned short)-559014399;
 	rTSQMG9jc	=	(unsigned short)-123987900;
-	HisuZ1Tn	=	761447537;
+	HisuZ1Tn	=	(unsigned short)761447537;
 	lgH7FK	=	-2026957172;
 	wii9C2	=	-692476027;
 	HVr10	=	-726995359;
@@ -68,7 +68,7 @@ int main() {
 	WcRn	=	(unsigned short)-103602187;
 	eri4N7ko	=	(unsigned short)-1186899969;
 	SQI2g0lS8	=	(unsigned short)-1734199649;
-	Z76k_A2J_	=	346730950;
+	Z76k_A2J_	=	(unsigned short)346730950;
 	T0gzp7bnl	=	-2006039250;
 	BaX13	=	-335425650;
 	i6_lHRZk	=	895445722;
@@ -78,7 +78,7 @@ int main() {
 	hkha1	=	(unsigned short)-1625717713;
 	NCO9Rz7RDb	=	(unsigned short)-647662377;
 	jBFx	=	(unsigned short)-1839163037;
-	JrIvSreY	=	-754685068;
+	JrIvSreY	=	(unsigned short)-754685068;
 	ZshXX_	=	-760534980;
 	lrz__YjT	=	-1896342821;
 	oGTC20JjV	=	-2116380059;
@@ -88,7 +88,7 @@ int main() {
 	IGoZUCd	=	(unsigned short)-61552867;
 	eJE63_b	=	(unsigned short)326305653;
 	VCuM3g	=	(unsigned short)2033577319;
-	g4b7l3E3	=	-1918695512;
+	g4b7l3E3	=	(unsigned short)-1918695512;
 	cdpA3vDt9	=	859574619;
 	kkVAuL4Y	=	1961959767;
 	OgAk_4v	=	-968854958;
