+++ main1161.cpp
@@ -28,7 +28,7 @@ int main(void)
 	short int	PBnq	=	(short)-972497304;
 	short int	J5HzP30m	=	(short)569292447;
 	long	h0M_GuU4	=	180171993;
-	unsigned short int	em47	=	-1245094068;
+	unsigned short int	em47	=	(unsigned short)-1245094068;
     unsigned short    LRlX13gMo    =    -1550964817;
 	short int	Ufxw	=	(short)1405172730;
 	char	t3v3Xb59	=	(char)336561102;
